requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.2.0+cu121
torchvision==0.17.0+cu121

timm==1.0.12
diffusers==0.32.1
accelerate>=0.26.0
torchdiffeq>=0.2.3
pytorch_fid>=0.3.0
tensorboard==2.16.2
omegaconf==2.3.0
einops>=0.7.0
fairscale>=0.4.13
safetensors>=0.4.2

pillow>=10.0.0
opencv-python>=4.8.0

matplotlib>=3.7.0
tqdm>=4.65.0

torchmetrics>=1.0.0

numpy>=1.24.0
pyyaml>=6.0
requests>=2.31.0